apps/heater-api/pyproject.toml
[project]
name = "heater-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi-camelcase>=2.0.0",
  "fastapi[standard]>=0.116.1",
  "lgg_client",
  "opentelemetry-api>=1.38.0",
  "opentelemetry-exporter-otlp-proto-grpc>=1.38.0",
  "opentelemetry-instrumentation-fastapi>=0.59b0",
  "opentelemetry-sdk>=1.38.0",
  "pyyaml>=6.0.3",
  "simulator_client",
]

[dependency-groups]
dev = ["autopep8>=2.3.2"]

[tool.uv.sources]
lgg_client = { workspace = true }
simulator_client = { workspace = true }
